requirements.txt
numpy == 1.14.1
scipy == 1.0.0
scikit-learn == 0.19.1
matplotlib == 2.1.2
pytorch == 0.3.0
